refactor: unified directory components to src/components
