+++ pages/_app.tsx
@@ -1,6 +1,6 @@
 import '../styles/tailwind.css';
 import type { AppProps } from 'next/app';
-import Layout from '../components/layout';
+import Layout from '../src/components/layout';
 import { GoogleAnalytics } from '../src/components/GoogleAnalytics';
 
 function MyApp({ Component, pageProps }: AppProps) {
